+++ shop_service/src/main/java/com/luckytree/shop_service/common/enums/ShopStatus.java
@@ -1,0 +1,5 @@
+package com.luckytree.shop_service.common.enums;
+
+public enum ShopStatus {
+    ENABLE, DISABLE
+}
