+++ shop_service/src/main/java/com/luckytree/shop_service/shop/adapter/data/ShopRequest.java
@@ -2,6 +2,7 @@ package com.luckytree.shop_service.shop.adapter.data;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.datatype.jsr310.deser.LocalTimeDeserializer;
+import com.luckytree.shop_service.common.enums.Category;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
