+++ shop_service/src/main/java/com/luckytree/shop_service/shop/adapter/persistence/ShopAdapter.java
@@ -1,14 +1,13 @@
 package com.luckytree.shop_service.shop.adapter.persistence;
 
 import com.luckytree.shop_service.shop.adapter.data.BookmarkDto;
-import com.luckytree.shop_service.shop.adapter.data.Category;
+import com.luckytree.shop_service.common.enums.Category;
 import com.luckytree.shop_service.shop.adapter.data.ShopRequest;
 import com.luckytree.shop_service.shop.application.port.outgoing.ShopPort;
-import com.luckytree.shop_service.shop.domain.Hashtag;
-import com.luckytree.shop_service.shop.domain.ShopDetail;
-import com.luckytree.shop_service.shop.adapter.data.ShopDetailDto;
-import com.luckytree.shop_service.shop.domain.ShopStatus;
-import com.luckytree.shop_service.shop.domain.ShopSummary;
+import com.luckytree.shop_service.common.enums.Hashtag;
+import com.luckytree.shop_service.shop.adapter.data.ShopDetail;
+import com.luckytree.shop_service.common.enums.ShopStatus;
+import com.luckytree.shop_service.shop.adapter.data.ShopSummary;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import org.webjars.NotFoundException;
