+++ shop_service/src/main/java/com/luckytree/shop_service/shop/application/port/incoming/ShopUseCase.java
@@ -1,11 +1,11 @@
 package com.luckytree.shop_service.shop.application.port.incoming;
 
-import com.luckytree.shop_service.shop.adapter.data.Category;
+import com.luckytree.shop_service.common.enums.Category;
 import com.luckytree.shop_service.shop.adapter.data.MyBookmarksDto;
 import com.luckytree.shop_service.shop.adapter.data.ShopRequest;
-import com.luckytree.shop_service.shop.domain.Hashtag;
-import com.luckytree.shop_service.shop.domain.ShopDetail;
-import com.luckytree.shop_service.shop.domain.ShopSummary;
+import com.luckytree.shop_service.common.enums.Hashtag;
+import com.luckytree.shop_service.shop.adapter.data.ShopDetail;
+import com.luckytree.shop_service.shop.adapter.data.ShopSummary;
 
 import java.util.List;
