+++ shop_service/src/main/java/com/luckytree/shop_service/shop/application/service/ShopService.java
@@ -1,16 +1,16 @@
 package com.luckytree.shop_service.shop.application.service;
 
 import com.luckytree.shop_service.shop.adapter.data.BookmarkDto;
-import com.luckytree.shop_service.shop.adapter.data.Category;
+import com.luckytree.shop_service.common.enums.Category;
 import com.luckytree.shop_service.shop.adapter.data.MyBookmarksDto;
 import com.luckytree.shop_service.shop.adapter.data.ShopRequest;
 import com.luckytree.shop_service.shop.adapter.persistence.ShopEntity;
 import com.luckytree.shop_service.shop.application.port.incoming.RemoveRequestForm;
 import com.luckytree.shop_service.shop.application.port.incoming.ShopUseCase;
 import com.luckytree.shop_service.shop.application.port.outgoing.ShopPort;
-import com.luckytree.shop_service.shop.domain.Hashtag;
-import com.luckytree.shop_service.shop.domain.ShopDetail;
-import com.luckytree.shop_service.shop.domain.ShopSummary;
+import com.luckytree.shop_service.common.enums.Hashtag;
+import com.luckytree.shop_service.shop.adapter.data.ShopDetail;
+import com.luckytree.shop_service.shop.adapter.data.ShopSummary;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
