+++ shop_service/src/main/java/com/luckytree/shop_service/shop/domain/Shop.java
@@ -1,5 +1,4 @@
 package com.luckytree.shop_service.shop.domain;
 
-public enum ShopStatus {
-    ENABLE, DISABLE
+public class Shop {
 }
